Upstream (#1)
* cmake_tool: Add Rocketchip platform_strings

This commit allows sel4test to be compiled with
-DPLATFORM=rocketchip-base and DPLATFORM=rocketchip-zcu102.

Signed-off-by: Robbie VanVossen <robert.vanvossen@dornerworks.com>

* cmake_tool: Build binaries for rocket chip plats

Signed-off-by: Robbie VanVossen <robert.vanvossen@dornerworks.com>

* github: bump GitHub action versions

Update to current node16 versions. The old node12 actions are
deprecated and will stop working.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>

* CI: simulation and HW build with LLVM on RISCV

Signed-off-by: Axel Heider <axelheider@gmx.de>

* CI: bring back compiler to matrix

This was deleted by accident in commit 75fc8008

Signed-off-by: Axel Heider <axelheider@gmx.de>

* lld: Enable building C projects without crt files

crt[begin|end].o files are used for C++ projects. While they usually
exist in GCC, they may or may not exist in LLVM builds that don't
rely on GCC toolchain, and typically require a separate (optional)
compiler-rt/cxx library to build.

This commit allows building projects without crt[begin|end].o files
if they do not exist.

Signed-off-by: Hesham Almatary <hesham.almatary@cl.cam.ac.uk>

* github: bump actions to node20

GitHub has started issuing warnings for node16 actions.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>

* elfloader: Add driver callback for non-boot core

This is to support core-specific driver init handling.

Signed-off-by: Kent McLeod <kent@kry10.com>

* elfloader: Add generic-timer driver

This driver just sets CNTVOFF_El2 to 0 if possible.
This can only be done if the elfloader is in hyp-mode, otherwise nothing
happens.

Signed-off-by: Kent McLeod <kent@kry10.com>

* elfloader: Replace calls to reset_cntvoff

These calls can now be implemented via binding the /timer driver in the
elfloader's device tree configuration (located in the kernel's platform
directory).

Signed-off-by: Kent McLeod <kent@kry10.com>

* elfloader: Only include driver source on arm

RISC-V platforms don't use any extension drivers currently so it's not
useful to compile and link them in when not building for arm.

Signed-off-by: Kent McLeod <kent@kry10.com>

* elfloader: Remove non_boot_init() interface

This was previously added to handle per-core initialization of the
generic timer. Now that a generic driver mechanism is used to perform
this initialization the old interface is redundant.

Signed-off-by: Kent McLeod <kent@kry10.com>

* elfloader: Check return code of device init

Device initialization returns early if an error is encountered. Handle
this error by printing an Error message and aborting the booting
process.

Signed-off-by: Kent McLeod <kent@kry10.com>

* CI: the seL4/machine_queue repo is public now

Signed-off-by: Axel Heider <axelheider@gmx.de>

* CI: cancel older concurrent PR runs

- Remove the space in the workflow name to ensure there are no side
  effects when using it as an identifier.
- Align workflow naming
- Act on PRs only, not on upstreaming.

Signed-off-by: Axel Heider <axelheider@gmx.de>

* misc: remove deprecated astyle option

Release 3.0 of astyle replaced this option with
'max-continuation-indent'. The replaced option is already in
the astylerc so we just need to remove the option here.

Further information can be found in the release notes:
https://astyle.sourceforge.net/news.html

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>

* Add support for Pine64 Star64

Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>

* cmake: Support binary build for imx8mp-evk

The imx8mp-evk as like with the other imx8m-* platforms require the
resulting image to just be a binary and not an ELF or other special
formats.

Signed-off-by: Damon Lee <damon@kry10.com>

* misc: check that astyle exists

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>

* Shift 32-bit image start address on imx8mm

The port of the hardware debug API to aarch64 has increased the size
of some debug images to the point that the address range the image is
loaded at overlaps with the address range the elf-loader relocates to
and expects to run at. This was solved on aarch64 by shifting the uboot
load address to 0x42000000. We do not differentiate between aarch32 and
aarch64 images in uboot, and the aarch32 elfloader is not relocated, so
the configuration in this commit must be changed to have the aarch32
elfloader expect to be executed from the new load address.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>

* cmake: use quotes

Avoid odd error messages in case TRIPLE is empty.

Signed-off-by: Axel Heider <axel.heider@codasip.com>

* trivial: style fix for autopep8 version bump

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>

* elfloader: initialise stack allocated struct

Only results in a compilation error on newer GCC versions.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>

* cmake_tool: add ELF loader setting for zynqmq 32

Enable the ELF loader for 32-bit images on zynqmq platforms.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>

---------

Signed-off-by: Robbie VanVossen <robert.vanvossen@dornerworks.com>
Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>
Signed-off-by: Axel Heider <axelheider@gmx.de>
Signed-off-by: Hesham Almatary <hesham.almatary@cl.cam.ac.uk>
Signed-off-by: Kent McLeod <kent@kry10.com>
Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>
Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>
Signed-off-by: Damon Lee <damon@kry10.com>
Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>
Signed-off-by: Axel Heider <axel.heider@codasip.com>
Co-authored-by: Robbie VanVossen <robert.vanvossen@dornerworks.com>
Co-authored-by: Gerwin Klein <gerwin.klein@proofcraft.systems>
Co-authored-by: Axel Heider <axelheider@gmx.de>
Co-authored-by: Hesham Almatary <hesham.almatary@cl.cam.ac.uk>
Co-authored-by: Kent McLeod <kent@kry10.com>
Co-authored-by: Ivan Velickovic <i.velickovic@unsw.edu.au>
Co-authored-by: Damon Lee <damon@kry10.com>
Co-authored-by: Alwin Joshy <joshyalwin@gmail.com>
Co-authored-by: Axel Heider <axel.heider@codasip.com>